+++ src/main/java/firok/tiths/TinkersThings.java
@@ -29,7 +29,7 @@ public class TinkersThings
 
 	public static final String MOD_ID = "tiths";
 	public static final String MOD_NAME = "TinkersThings";
-	public static final String VERSION = "1.12.2-0.2.6.0";
+	public static final String VERSION = "1.12.2-0.2.7.0";
 
 
 	{