+++ src/main/java/firok/tiths/common/DamageSources.java
@@ -2,6 +2,7 @@ package firok.tiths.common;
 
 import net.minecraft.util.DamageSource;
 
+@SuppressWarnings("all")
 public class DamageSources
 {
 	// 精神伤害
@@ -10,4 +11,6 @@ public class DamageSources
 			.setDamageBypassesArmor();
 	// 双折伤害
 	public static final DamageSource BirefringentDamage=new DamageSource("birefringent");
+	// 热释电伤害
+	public static final DamageSource PyroelectricDamage=new DamageSource("pyroelectric");
 }