+++ src/main/java/firok/tiths/common/TileEntities.java
@@ -1,0 +1,9 @@
+package firok.tiths.common;
+
+import firok.tiths.block.logic.BlockLogicSearing;
+
+// 方块实体
+public class TileEntities
+{
+	public static BlockLogicSearing logic_searing;
+}